split repositories into interface and implementation
